refactor: 컴포넌트 선언 방식 변경, index 컴포넌트 추가
